fix: rename elements dir to selectors
