+++ cypress/integration/ui/item_context_menu/item_context_menu.js
@@ -6,7 +6,7 @@ import {
     itemDetailsSel,
     clickMenuButton,
     getDashboardItem,
-} from '../../../elements/dashboardItem'
+} from '../../../selectors/dashboardItem'
 import { dashboards } from '../../../assets/backends'
 
 // these tests being run on the "Delivery" dashboard
