+++ cypress/selectors/titleBar.js
